backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
<<<<<<< HEAD
djangorestframework==3.14.0
=======
>>>>>>> 7bd8d60caea551423507fb6cadb6b5d26120506a
idna==3.6
joblib==1.3.2
numpy==1.26.2
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
urllib3==2.1.0
